refactor: make add venue its own component and switch it over to using api constants for consistency
