+++ src/utils/constants.mjs
@@ -1,3 +1,3 @@
 export const BASE_URL = `https://v2.api.noroff.dev/holidaze`;
 
-export const ALL_USERS = `${BASE_URL}/venues`;
+export const VENUES = `${BASE_URL}/venues`;
